Implement "check" feature.
This adds a system for voting that will also be helpful for other
features like "Reveal" or "Give up". When a user wishes to check an
answer (or the entire grid), it's put up for a vote amongst the other
participants. When everyone agrees, the last user does the check. A vote
has 30s to succeed.

The wave state records a set of (x, y, letter) tuples which have been
shown to be incorrect. This lets the UI show either an "X" for an
actively wrong answer or a "\" dog-ear for an answer which was
previously shown to be wrong.

The voting UI is fairly ugly, but it works!